Split devices plugin: devices: These only handle udev signals, and distribute these. (Hidden to final user) mtp: Handle mtp devices when devices-plugin inform this. removable: Handle removable devices when devices-plugin inform this. cdrom: The same, but detect cdrom added and suggest append it
