+++ configure.ac
@@ -261,6 +261,8 @@ AC_CONFIG_FILES([plugins/gnome-media-keys/Makefile])
 AC_CONFIG_FILES([plugins/song-info/Makefile])
 AC_CONFIG_FILES([plugins/notify/Makefile])
 AC_CONFIG_FILES([plugins/devices/Makefile])
+AC_CONFIG_FILES([plugins/mtp/Makefile])
+AC_CONFIG_FILES([plugins/removable-media/Makefile])
 AC_CONFIG_FILES([plugins/tunein/Makefile])
 AC_CONFIG_FILES([plugins/dlna/Makefile])
 AC_CONFIG_FILES([plugins/cdrom/Makefile])
