+++ plugins/Makefile.am
@@ -25,8 +25,13 @@ if HAVE_LIBNOTIFY
 SUBDIRS += notify
 endif
 
-if HAVE_LIBMTP
+if HAVE_GUDEV
 SUBDIRS += devices
+SUBDIRS += removable-media
+endif
+
+if HAVE_LIBMTP
+SUBDIRS += mtp
 endif
 
 if HAVE_LIBSOUP
