+++ plugins/devices/devices.plugin
@@ -6,4 +6,4 @@ Authors=Matias De lellis <mati86dl@gmail.com>
 Copyright=Copyright © 2014 Matias De lellis
 Website=https://github.com/matiasdelellis/pragha
 Help=https://github.com/matiasdelellis/pragha
-
+Hidden=true
