+++ plugins/removable-media/Makefile.am
@@ -1,21 +1,18 @@
-plugindir = $(libdir)/pragha/plugins/devices
+plugindir = $(libdir)/pragha/plugins/removable
 
 AM_CPPFLAGS = \
 	-I$(top_srcdir) 	\
 	$(LIBPEAS_CFLAGS)		\
 	$(LIBPEAS_GTK_CFLAGS)
 
-plugin_LTLIBRARIES = libdevices.la
+plugin_LTLIBRARIES = libremovable.la
 
-libdevices_la_SOURCES = \
-	pragha-devices-plugin.c \
-	pragha-devices-plugin.h \
-	pragha-device-client.c \
-	pragha-device-client.h
+libremovable_la_SOURCES = \
+	pragha-devices-removable.c
 
-libdevices_la_LDFLAGS = $(PLUGIN_LIBTOOL_FLAGS)
+libremovable_la_LDFLAGS = $(PLUGIN_LIBTOOL_FLAGS)
 
-libdevices_la_CFLAGS =						\
+libremovable_la_CFLAGS =						\
 	$(GSTREAMER_CFLAGS)						\
 	$(GSTREAMER_BASE_CFLAGS)					\
 	$(GLIB_CFLAGS) 							\
@@ -29,7 +26,7 @@ libdevices_la_CFLAGS =						\
 	-I$(top_srcdir)/src/include 					\
 	-Wall
 
-libdevices_la_LIBADD =							\
+libremovable_la_LIBADD =							\
 	$(GSTREAMER_LIBS)						\
 	$(GLIB_LIBS) 							\
 	$(GIO_LIBS) 							\
@@ -41,15 +38,15 @@ libdevices_la_LIBADD =							\
 	$(GUDEV_LIBS)
 
 if HAVE_GSTREAMER_AUDIO
-libdevices_la_CFLAGS += $(GSTREAMER_AUDIO_CFLAGS)
-libdevices_la_LIBADD += $(GSTREAMER_AUDIO_LIBS)
+libremovable_la_CFLAGS += $(GSTREAMER_AUDIO_CFLAGS)
+libremovable_la_LIBADD += $(GSTREAMER_AUDIO_LIBS)
 endif
 
 if HAVE_LIBXFCE4UI
-libdevices_la_CFLAGS += $(LIBXFCE4UI_CFLAGS)
-libdevices_la_LIBADD += $(LIBXFCE4UI_LIBS)
+libremovable_la_CFLAGS += $(LIBXFCE4UI_CFLAGS)
+libremovable_la_LIBADD += $(LIBXFCE4UI_LIBS)
 endif
 
-plugin_DATA = devices.plugin
+plugin_DATA = removable.plugin
 
 EXTRA_DIST = $(plugin_DATA)
