+++ plugins/removable-media/removable.plugin
@@ -1,7 +1,7 @@
 [Plugin]
-Module=cdrom
-Name=Play cdrom
-Description=Play audios cds
+Module=removable
+Name=Removable media
+Description=Detect removable media and scan it
 Authors=Matias De lellis <mati86dl@gmail.com>
 Copyright=Copyright © 2014 Matias De lellis
 Website=https://github.com/matiasdelellis/pragha
